Add test/example images
Images used for demonstration purposes. Read the full description in the
article published in the German IT journal "Web & Mobile Developer".

All images have been taken from Pexels [1] and are licensed under the
Creative Commons Zero (CC0) license. This means the pictures are
completely free to be used for personal and commercial purposes.

[1] https://www.pexels.com